FW-708. change uint16_t to indicate the length of packet.

+++ bsp/boards/openmote-b/spi.c
@@ -20,10 +20,10 @@
 
 
 //=========================== defines =========================================
-#define SPI_PIN_SSI_CLK             GPIO_PIN_2      //    CLK       
-#define SPI_PIN_SSI_FSS             GPIO_PIN_3      //    CSn       
-#define SPI_PIN_SSI_RX              GPIO_PIN_4      //    MISO      
-#define SPI_PIN_SSI_TX              GPIO_PIN_5      //    MOSI     
+#define SPI_PIN_SSI_CLK             GPIO_PIN_2      //    CLK
+#define SPI_PIN_SSI_FSS             GPIO_PIN_3      //    CSn
+#define SPI_PIN_SSI_RX              GPIO_PIN_4      //    MISO
+#define SPI_PIN_SSI_TX              GPIO_PIN_5      //    MOSI
 #define SPI_GPIO_SSI_BASE           GPIO_A_BASE
 
 //=========================== variables =======================================
@@ -55,36 +55,36 @@ static void enableInterrupts(void);
 
 void spi_init() {
     // clear variables
-    memset(&spi_vars,0,sizeof(spi_vars_t));  
-    
+    memset(&spi_vars,0,sizeof(spi_vars_t));
+
     // set the clk miso and cs pins as output
     GPIOPinTypeGPIOOutput(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_CLK);
     GPIOPinTypeGPIOOutput(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_TX);
     GPIOPinTypeGPIOOutput(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_FSS);
-    
+
     //set cs to high
     GPIOPinWrite(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_FSS, SPI_PIN_SSI_FSS);
     //set pins to low
     GPIOPinWrite(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_TX, 0);
     GPIOPinWrite(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_CLK, 0);
-    
+
     SysCtrlPeripheralEnable(SYS_CTRL_PERIPH_SSI0);
     SysCtrlPeripheralSleepEnable(SYS_CTRL_PERIPH_SSI0);
     SysCtrlPeripheralDeepSleepDisable(SYS_CTRL_PERIPH_SSI0);
-    
+
     SSIDisable(SSI0_BASE);
     SSIClockSourceSet(SSI0_BASE, SSI_CLOCK_PIOSC);
-  
-    IOCPinConfigPeriphOutput(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_CLK, IOC_MUX_OUT_SEL_SSI0_CLKOUT);    
-    IOCPinConfigPeriphOutput(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_TX, IOC_MUX_OUT_SEL_SSI0_TXD);    
-    IOCPinConfigPeriphInput(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_RX, IOC_SSIRXD_SSI0);    
-    
+
+    IOCPinConfigPeriphOutput(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_CLK, IOC_MUX_OUT_SEL_SSI0_CLKOUT);
+    IOCPinConfigPeriphOutput(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_TX, IOC_MUX_OUT_SEL_SSI0_TXD);
+    IOCPinConfigPeriphInput(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_RX, IOC_SSIRXD_SSI0);
+
     GPIOPinTypeSSI(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_CLK );
     GPIOPinTypeSSI(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_RX );
     GPIOPinTypeSSI(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_TX );
-    
+
     SSIConfigSetExpClk(SSI0_BASE, SysCtrlIOClockGet(), SSI_FRF_MOTO_MODE_0, SSI_MODE_MASTER, /*SysCtrlIOClockGet()/2*/16000000, 8);
-    
+
     // Enable the SSI0 module.
     SSIEnable(SSI0_BASE);
 }
@@ -96,10 +96,10 @@ void spi_setCb(spi_cbt cb) {
 #endif
 
 void    spi_txrx(uint8_t*     bufTx,
-                 uint8_t      lenbufTx,
+                 uint16_t     lenbufTx,
                  spi_return_t returnType,
                  uint8_t*     bufRx,
-                 uint8_t      maxLenBufRx,
+                 uint16_t     maxLenBufRx,
                  spi_first_t  isFirst,
                  spi_last_t   isLast) {
 
@@ -114,15 +114,15 @@ void    spi_txrx(uint8_t*     bufTx,
     spi_vars.maxRxBytes       =  maxLenBufRx;
     spi_vars.isFirst          =  isFirst;
     spi_vars.isLast           =  isLast;
-   
+
     // SPI is now busy
     spi_vars.busy             =  1;
-    
+
     // lower CS signal to have slave listening
     if (spi_vars.isFirst==SPI_FIRST) {
        GPIOPinWrite(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_FSS, 0);
     }
-    
+
     for ( i =  0; i < lenbufTx; i++)
     {
         // Push a byte
@@ -133,12 +133,12 @@ void    spi_txrx(uint8_t*     bufTx,
 
         // Read a byte
         SSIDataGet(SSI0_BASE, &data);
-        
+
         // Store the result
         spi_vars.pNextRxByte[i] = (uint8_t)(data & 0xFF);
         // one byte less to go
      }
-   
+
      if (spi_vars.isLast==SPI_LAST) {
         GPIOPinWrite(SPI_GPIO_SSI_BASE, SPI_PIN_SSI_FSS, SPI_PIN_SSI_FSS);
      }
@@ -171,32 +171,32 @@ port_INLINE void disableInterrupts(void)
 //=========================== interrupt handlers ==============================
 
 kick_scheduler_t spi_isr() {
-#ifdef SPI_IN_INTERRUPT_MODE  
+#ifdef SPI_IN_INTERRUPT_MODE
     uint32_t data;
     // save the byte just received in the RX buffer
     status = SSIIntStatus(SSI0_BASE, true);
 
     // Clear SPI interrupt in the NVIC
     IntPendClear(INT_SSI0);
-    
+
     SSIDataGet(SSI0_BASE, &data);
 
     // Store the result
     spi_vars.pNextRxByte = (uint8_t)(data & 0xFF);
-       
+
     // one byte less to go
     spi_vars.pNextTxByte++;
     spi_vars.pNextRxByte++;
     spi_vars.txBytesLeft--;
-   
+
     if (spi_vars.txBytesLeft>0) {
         // write next byte to TX buffer
         SSIDataPut(SSI0_BASE, *spi_vars.pNextTxByte);
-          
+
     } else {
         // SPI is not busy anymore
         spi_vars.busy             =  0;
-      
+
         // SPI is done!
         if (spi_vars.callback!=NULL) {
            // call the callback
@@ -205,7 +205,7 @@ kick_scheduler_t spi_isr() {
             return KICK_SCHEDULER;
         }
     }
-    
+
 #endif
     return DO_NOT_KICK_SCHEDULER;
 }
